add initial take of bloom filter
this is not the form it will be kept in. stability of filter between
snapshots as long as size remains same results in same object with
same hash being emitted more than once, which is not a desirable
property for the storage backend to deal with and fails to "age out"
false positives. committed just as baseline for coming changes.