add beginnings of a prune subcommand
the concept here is to take a set of bloom filters for objects to
keep, and select all non-matching objects for deletion. significant
work still needs to be done designing an interface that's easy to use
and not a footgun. the current code is more just a proof of concept
that the bloom filter works as intended for garbage collection.