remove 'drops' gc support
this or something like it may be re-added at some point, but it proved
to be a bad idea because it doesn't admit abandoning any backups
except the oldest. in particular it does not support retaining
frequent backups for some range into the past, then only infrequent
backups further back.

the replacement, bloom filter, does not have this problem, but does
involve storage cost proportional to the number of object blobs for
each snapshot. this may be addressed in the future with differential
bloom filters, and can be mitigated somewhat by only including bloom
filters with snapshots that will be kept longer than the gc interval.