prefix block contents with a type identifier for hashing
mainly, this will allow both block and inode label hashes to be used
as keys in the index without a concern that a block could contain data
whose form is a valid inode label, thereby colliding.

it also serves a minor secondary purpose: making it so block hashes
from an accidentally exposed index are not raw sha3 hashes of data
that are likely to be present in search engines for reversing the
hash. of course this does not prevent someone from building a reverse
mapping table with hashes of the "blk\0"-prefixed contents of
well-known files, but it still seems to have some privacy value.