switch working index from in-memory AVL tree to on-disk prefix tree
the in-memory AVL tree is way too large for backing up large
filesystems, especially from a 32-bit host where address space will be
exhausted. replace it with 'flatmap', a minimalist prefix tree based
map/dictionary data structure designed for insert- and query-only
loads. it's walkable in-place on disk. currently it's only constructed
from the text based index as a working state, but after cleaning up a
few more things, the intent is that it become the on-disk index
format. this will greatly reduce startup time and avoid needing disk
space to store both versions of the index (text and working flatmap).

iteration for bloom filter generation is not yet present, and this is
somewhat messy still, but I'm committing it as a checkpoint for
comparing further work done.