suffix hash keys with a type only for ino.block
since commit f4271e4892 inode label hashes can no longer clash with
blocks, so they can live in a common namespace. by making the
ino.block label keys one byte longer, we'll be able to filter them as
needed in the tree walk to build the bloom filter.

this is not the desired long-term solution. I want to get rid of the
ino.block keys entirely and instead store the list of blocks with the
main inode index entry. the only reasons that isn't done now are (1)
it didn't fit well in the text based index format built on lines with
key/value pairs, and (2) the ino.block index records were added late
in the design after I realized there was no way to maintain references
to the blocks from previous backups without them. this will be
revisited after moving to the native flatmap index format.