make flatmap_set report and fail on write errors
the order of operations is such that the file is still valid, just
with possibly wasted space, until the very last operation succeeds. an
eventual enhancement could be to ftruncate back to the original length
on any failure, but in practice we'll be throwing away the whole file
on error anyway for now.