rework flatmap iterator function
allow any possible keylength up to the limit (255). this is a
sufficient depth limit to prevent unsafe recursion anyway, so drop
depth limit from external interface. since we don't know (don't store)
the lengths of values, don't try to load a value for the callback;
just let it use the new flatmap_read function to read a value of
whatever length it wants. pass the flatmap context pointer back to the
callback so this works.