prune: overhaul to make it viable for real use
accept a list of summary files and search for objects relative to the
first, applying bloom filters from each. error out if anything goes
wrong. produce the list of objects to delete by default (e.g. for
piping to xargs). allow the inverse operation as well, for subsetting
objects to keep (e.g. archiving a single backup snapshot or some
subset in isolation).