backup: add verbose output option
so far this is very primitive and just prints the pathnames that have
changed. it could be enhanced later to show number of changed blocks,
size of new data stored, etc. and perhaps even to show object hashes
involved.

maintenance of a complete pathname buffer is added as part of this
change, and it is maintained independently of the verbose option. this
is groundwork for adding support for file exclusion filters, which are
hard to do without an expanded-out pathname.