backup: add support for exclusion patterns
patterns are read from an exclude file in the index directory. its
format is intended to be a subset of what's accepted in .gitignore,
with the main (only?) omission being inverted (exception) rules using
a leading !, which are not supported.

matching is implemented by converting the glob-style patterns to regex
(POSIX ERE) format, concatenating them with the regex disjunction (|)
operator, and compiling the whole thing as a single regex for
efficient application. this means there should be effectively no
penalty for using a very large number of patterns, if needed.

if inverted rules are eventually wanted, they could be added by
starting a new regex each time the inversion status flips.