backup: fix regression in verbose output
changed made in commit a1c88f0bd9b4c58d0cec4e6b1e285de806d80583 failed
to truncate the pathname string when leaving a completed directory. at
this point the string was used only for verbose output, not for
matching, so it did not cause wrong behavior but produced confusing
verbose output (showing the pathname of the last-processed file in
place of the pathname for the directory).