backup: read symlink contents before committing to store the symlink
when racing against a changing filesystem, reading the link may fail,
and this is okay. but if we've already started storing the inode, it's
not easy to handle that failure gracefully. previously, the
possibility of failure was ignored, leaking uninitialized memory into
the backed-up symlink contents.

move the readlink operation to the point of open/stat, so that it can
cause the inode to be skipped and warned about like other failures to
open.