restore: add option for recreating localindex for incremental backup
after restoring from a backup, continued usage of that backup with
incremental updates requires having the localindex for it. backing up
the localindex is not a good idea, as it's large and rewritten from
scratch each time rather than just growing by append, requiring a new
full copy to be stored, and the index available at backup time is for
the previous backup, not the new one. using the old one could even
cause data loss if the corresponding backup snapshot had been pruned.
moreover, the inode numbers in the index will no longer be valid after
restoring from the backup to a new filesystem, so only the block part
of the index would be useful anyway.

so, instead of including the index in backups, regenerate it as part
of the restore operation, filling in the new inode numbers obtained
during restore. these may not be useful if the restore is being staged
on non-final media, but the block records will be usable regardless.
the restored index timestamp is set to the time after completing
restore, so that inodes which do not change after restore will be
treated as up-to-date in the backup.

presently there is no error checking or reporting during generation of
the new index. if writing fails, the index will just be incomplete,
resulting in redundant storage on next backup. it would probably be
nice to add a warning about this.

also, the restored index lacks device mapping; this means files
restored to non-backup-root devices will only have numeric device
numbers in the index, likely preventing the inode records (but not
block records) from being reused. this could be remedied by adding
device mappings as an input to the restore operation, or by making
backup fallback to looking up numeric device numbers in the index if
the dev/inode is not found by device label.