cull: new utility for managing backup retention
this is still a work in progress. the cull subcommand takes a list of
backup summary files, extracts their timestamps, and uses those, along
with policy options expressed on the command line to cull old backups.

the -r option allows requesting that all backups within a particular
range of the latest timestamp be kept (default: 86400 seconds).

the -d, -w, -m, and -y options allow requesting that the latest backup
for a number of days, weeks, months, or years be kept. there is
presently no control over the cutoff for what defines a day, week,
month, or year boundary; it's always end of the last day of the period
in UTC.

names of files to be culled (removed) are printed to stdout. this
behavior can be inverted with -v. like with the purge subcommand, no
files are actually removed; the output can be piped to xargs to
perform removal or to stage them in a different directory for removal.

backup summaries without bloom filters are never considered candidates
for keeping, since it's impossible to determine what objects they
depend on.