backup: don't produce error output for unix sockets
open fails with ENXIO for unix sockets (as well as for some odd cases
with device nodes, which should not be being backed up anyway).
counting this as an error (presently ignored) or putting error
messages in logs may give a wrong impression that the backup failed,
so silently ignore it.

arguably we should switch to performing fstatat before open, skipping
open except for regular files, and only redoing fstat if an actual
file was opened; this would avoid doing weird stuff with
non-regular-files in all non-race conditions.